Update README.md
Apart of a project in which I created a model for climbing the Empire State building. There are 60 floors and and a dice role determines if you go up or down in floors. The result of each dice roll is presented using a histogram to show that in a simulation of events, the rest will be presented in the form of a bell shaped curve (Normal Distribution)